requirements.txt
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.113
botocore==1.29.113
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
cx-Oracle==8.3.0
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.0
fastapi-cache2==0.2.1
forex-python==1.8
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hiredis==2.2.2
idna==3.4
jmespath==1.0.1
multidict==6.0.4
#mysqlclient==2.1.1
openpyxl==3.1.2
pandas==2.0.3
passlib[bcrypt]==1.7.4
pendulum==2.1.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
pytzdata==2020.1
redis==4.5.4
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starkbank-ecdsa==2.2.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2
